Crash in com.apple.WebKit.WebContent at com.apple.JavaScriptCore: JSC::revertCall + 24 https://bugs.webkit.org/show_bug.cgi?id=145527
Reviewed by Filip Pizlo.

If a CallLinkInfo is GC'ed, we need to notify any PolymorphicCallNode's that reference it.
Added plumbling to clear the m_callLinkInfo of a PolymorphicCallNode when that CallLinkInfo
is going away.

* bytecode/CallLinkInfo.h:
(JSC::CallLinkInfo::~CallLinkInfo):
* jit/PolymorphicCallStubRoutine.cpp:
(JSC::PolymorphicCallNode::unlink):
(JSC::PolymorphicCallNode::clearCallLinkInfo):
(JSC::PolymorphicCallCase::dump):
(JSC::PolymorphicCallStubRoutine::edges):
(JSC::PolymorphicCallStubRoutine::clearCallNodesFor):
(JSC::PolymorphicCallStubRoutine::visitWeak):
* jit/PolymorphicCallStubRoutine.h:
(JSC::PolymorphicCallNode::hasCallLinkInfo):


Canonical link: https://commits.webkit.org/163569@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185084 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/JavaScriptCore/jit/PolymorphicCallStubRoutine.cpp
@@ -43,13 +43,23 @@ PolymorphicCallNode::~PolymorphicCallNode()
 
 void PolymorphicCallNode::unlink(RepatchBuffer& repatchBuffer)
 {
-    if (Options::showDisassembly())
-        dataLog("Unlinking polymorphic call at ", m_callLinkInfo->callReturnLocation, ", ", m_callLinkInfo->codeOrigin, "\n");
-    
-    m_callLinkInfo->unlink(repatchBuffer);
-    
+    if (m_callLinkInfo) {
+        if (Options::showDisassembly())
+            dataLog("Unlinking polymorphic call at ", m_callLinkInfo->callReturnLocation, ", ", m_callLinkInfo->codeOrigin, "\n");
+
+        m_callLinkInfo->unlink(repatchBuffer);
+    }
+
     if (isOnList())
         remove();
+}
+
+void PolymorphicCallNode::clearCallLinkInfo()
+{
+    if (Options::showDisassembly())
+        dataLog("Clearing call link info for polymorphic call at ", m_callLinkInfo->callReturnLocation, ", ", m_callLinkInfo->codeOrigin, "\n");
+
+    m_callLinkInfo = nullptr;
 }
 
 void PolymorphicCallCase::dump(PrintStream& out) const
@@ -95,6 +105,16 @@ CallEdgeList PolymorphicCallStubRoutine::edges() const
     for (size_t i = 0; i < m_variants.size(); ++i)
         result.append(CallEdge(CallVariant(m_variants[i].get()), m_fastCounts[i]));
     return result;
+}
+
+void PolymorphicCallStubRoutine::clearCallNodesFor(CallLinkInfo* info)
+{
+    for (Bag<PolymorphicCallNode>::iterator iter = m_callNodes.begin(); !!iter; ++iter) {
+        PolymorphicCallNode& node = **iter;
+        // All nodes should point to info, but okay to be a little paranoid.
+        if (node.hasCallLinkInfo(info))
+            node.clearCallLinkInfo();
+    }
 }
 
 bool PolymorphicCallStubRoutine::visitWeak(RepatchBuffer&)
